+++ scripts/azureml-assets/CHANGELOG.md
@@ -3,6 +3,14 @@
 
 ### 🐛 Bugs Fixed
 
+## 1.16.52 (2024-06-03)
+### 🚀 New Features
+- [#3000](https://github.com/Azure/azureml-assets/pull/3000) Allow use of azcopy's --overwrite flag
+
+## 1.16.51 (2024-05-30)
+### 🐛 Bugs Fixed
+- [#2991](https://github.com/Azure/azureml-assets/pull/2991) Generate SAS tokens for data assets
+
 ## 1.16.50 (2024-05-22)
 ### 🐛 Bugs Fixed
 - [#2956](https://github.com/Azure/azureml-assets/pull/2956) Update embeddings asset task type